dealer sales added discount and sr sales updated

+++ KGERP.Service/Implementation/Procurement/VMPurchaseOrder.cs
@@ -223,7 +223,7 @@ namespace KGERP.Service.Implementation.Procurement
         public int StockInfoTypeId { get; set; } // Come from StockFromTypeEnum
         public int? StockInfoId { get; set; } //CompanyWarehouse/ Depot/ Dealer/ Customer
         public string StockInfoName { get; set; }
-        public int SalePersonId { get; set; }
+        public long SalePersonId { get; set; }
         public string CompanyAddress { get; set; }
         public string CompanyEmail { get; set; }
         public string CompanyPhone { get; set; }
